+++ RELEASES.md
@@ -2,7 +2,11 @@
 
 ## Pending Release
 
+<<<<<<< Updated upstream
 - Removed deprecated flags `coreth-admin-api-enabled`, `coreth-admin-api-dir`, `tx-regossip-frequency`, `tx-lookup-limit`. Use `admin-api-enabled`, `admin-api-dir`, `regossip-frequency`, `transaction-history` instead.
+=======
+- Enabled RPC batch limits by default, and configurable with `batch-request-limit` and `batch-max-response-size`.
+>>>>>>> Stashed changes
 
 ## [v0.15.3](https://github.com/ava-labs/coreth/releases/tag/v0.15.3)